Rename service folder
