+++ labgrid-web-client/src/app/_services/place.service.ts
