+++ labgrid-web-client/src/app/_services/resource.service.ts
